opl Config move to opl namespace
