+++ sdl2-sonic-drivers/examples/opl-example.cpp
@@ -1,5 +1,5 @@
 #include <hardware/opl/OPL.hpp>
-#include <hardware/opl/scummvm/Config.hpp>
+#include <hardware/opl/Config.hpp>
 
 #include <hardware/opl/mame/MameOPL.hpp>
 
@@ -17,8 +17,8 @@
 
 
 using audio::scummvm::SdlMixerManager;
-using hardware::opl::scummvm::Config;
-using hardware::opl::scummvm::OplEmulator;
+using hardware::opl::Config;
+using hardware::opl::OplEmulator;
 using utils::delayMillis;
 
 using utils::FMoutput;
