+++ sdl2-sonic-drivers/src/hardware/opl/OPL.hpp
@@ -26,6 +26,7 @@ namespace hardware
             {
                 return _init;
             }
+
             /**
              * Initializes the OPL emulator.
              *
